requirements.txt
streamlit
tensorflow
Pillow
numpy
pickle-mixin